+++ src/Exception/Exception.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\Server\Status\Exception;
-
-interface Exception extends \Throwable
-{
-}